order-api/src/main/java/com/zerobase/cms/sevice/ProductService.java
package com.zerobase.cms.sevice;
import com.zerobase.cms.model.Product;
import com.zerobase.cms.model.product.AddProductForm;
import com.zerobase.cms.model.repository.ProductRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@RequiredArgsConstructor
public class ProductService {

	private final ProductRepository productRepository;

	@Transactional
	public Product addProduct(Long sellerId, AddProductForm form) {
		return productRepository.save(Product.of(sellerId, form));
	}
}